# Fix FA3 DeepSeek prefill performance regression