@zhyncs 2025-04-22
same https://github.com/sgl-project/sglang/pull/5623