@Fridge003 2025-04-22
@Alcanderian 
I feel the computation neglected some parts:
1. For MLA, the bmm for q_nope (128*128 to 128*512), and bmm after attn, they should also be added.
2. The precision of matrix multiplication is fp8, while the precision of attention is bf16